Image with vertical line down the center.
This image is important for calibration because for dome projection we assume
that the projector, mirror, dome, and observer are all aligned in the plane in
which this vertical line is projected.  If that is not the case no values of
the parameters will achieve the correct image inside the dome.