Rewrote find_center_pixels.
The old, recursive version was very slow to find the center pixels in the image
with 24 diamonds.  The new version is better in every way (New and Improved!).
It segments the image into a grid and creates an object for each cell.  As a
bonus this method made it easy to sort the resulting objects from right to left
and then top to bottom which is required for them to match up with the correct
image pixels (which are listed left to right and then top to bottom, since a
mirror is involved).