Added correction for camera distortion.
Extensive changes.  Made a fix to the code that generates diamonds for the
calibration images.  Changed the function that gets minimized for parameter estimation.  Added methods to estimate and remove camera distortion, though
it does not appear that this distortion is significant with the current dome
calibration image.  It occurs to me that the projector's optics will also
introduce distortion and the current calibration method does not correct for
this.