Streamlined the calibration process.
Parameters can now be passed to warp_image.py in a file as an argument.
The calibration image saved by dome_calibration.py now contains green
dots at the locations specified in the file saved using caldev.exe and
blue spots at the locations that map to the calibrated viewing
directions using the post calibration parameters.

These blue dots are found by generating a projector image from 22 or 37
copies of a small image with a blue dot in the center. This is much more
reliable that the optimization method being used before.